acivities folder created, project restructuring
