+++ App_mobile/app/src/main/AndroidManifest.xml
@@ -13,20 +13,20 @@
         android:theme="@style/Base.Theme.Libreria_In_47_App"
         tools:targetApi="31">
         <activity
-            android:name=".ContactActivity"
+            android:name=".activities.ContactActivity"
             android:exported="false" />
         <activity
-            android:name=".accountactivity"
+            android:name=".activities.accountactivity"
             android:exported="false" />
         <activity
-            android:name=".BookDetail"
+            android:name=".activities.BookDetail"
             android:exported="false" />
         <activity
-            android:name=".RegisterActivity"
+            android:name=".activities.RegisterActivity"
             android:exported="false"
             android:label="@string/title_activity_register" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -35,10 +35,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".AboutUsActivity"
+            android:name=".activities.AboutUsActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true" />
 
         <meta-data
