+++ App_mobile/app/src/main/java/com/example/libreria_in_47_app/activities/BookDetail.java
@@ -1,4 +1,4 @@
-package com.example.libreria_in_47_app;
+package com.example.libreria_in_47_app.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -8,6 +8,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
+
+import com.example.libreria_in_47_app.R;
 
 public class BookDetail extends AppCompatActivity {
     ImageView ivRegresar;
