+++ App_mobile/app/src/main/java/com/example/libreria_in_47_app/activities/ContactActivity.java
@@ -1,4 +1,4 @@
-package com.example.libreria_in_47_app;
+package com.example.libreria_in_47_app.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -9,6 +9,8 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.libreria_in_47_app.R;
 
 public class ContactActivity extends AppCompatActivity {
