+++ App_mobile/app/src/main/java/com/example/libreria_in_47_app/activities/RegisterActivity.java
@@ -1,13 +1,15 @@
-package com.example.libreria_in_47_app;
+package com.example.libreria_in_47_app.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
-import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.libreria_in_47_app.DataBaseSQLiteHelper;
+import com.example.libreria_in_47_app.R;
 
 public class RegisterActivity extends AppCompatActivity {
