+++ App_mobile/app/src/main/res/layout/activity_accountactivity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".accountactivity">
+    tools:context=".activities.accountactivity">
     <TextView
         android:id="@+id/textView4"
         android:layout_width="334dp"
