+++ App_mobile/app/src/main/res/layout/activity_contact.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactActivity">
+    tools:context=".activities.ContactActivity">
 
     <ImageView
         android:id="@+id/ivRegresar"
